@@ -13,11 +13,9 @@
     "source": "dist"
   },
   "scripts": {
-    "start": "webpack-dev-server --mode development",
-    "build": "webpack --mode production",
-    "lint": "eslint \"**/*.js\"",
-    "verify-i18n": "npm run i18n -- --check",
-    "i18n": "sync-i18n --files 'src/locales/*.json' --primary en --languages fr --space 2"
+    "start": "webpack-dev-server --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js",
+    "lint": "eslint \"**/*.js\""
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
@@ -25,16 +23,16 @@
     "@babel/preset-react": "^7.9.4",
     "babel-loader": "^8.1.0",
     "copy-webpack-plugin": "^8.1.1",
-    "eslint": "^6.8.0",
+    "eslint": "^8.1.0",
     "eslint-plugin-jasmine": "^4.1.1",
     "eslint-plugin-react": "^7.19.0",
     "html-loader": "^1.1.0",
     "html-webpack-plugin": "^4.2.0",
-    "i18next-json-sync": "^2.3.1",
     "webpack": "^5.31.2",
     "webpack-bundle-analyzer": "^4.4.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.2"
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.3.1",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
     "@material-ui/core": "^4.9.10",
@@ -46,7 +44,6 @@
     "babel-polyfill": "^6.26.0",
     "clsx": "^1.1.1",
     "i18next": "^19.4.3",
-    "i18next-browser-languagedetector": "^6.0.1",
     "lodash": "^4.17.20",
     "prop-types": "^15.7.2",
     "query-string": "^6.14.0",
