@@ -1,7 +1,10 @@
 #!/usr/bin/env node
-import cdk = require('@aws-cdk/core');
-import {StaticSite} from '../lib/cdk-stack';
-import {Tags} from '@aws-cdk/core';
+import * as cdk from '@aws-cdk/core';
+import * as s3deploy from "@aws-cdk/aws-s3-deployment";
+import * as origins from '@aws-cdk/aws-cloudfront-origins';
+import StaticDeployment from "../lib/StaticDeployment";
+import Notifications from "../lib/Notifications";
+import * as cloudfront from "@aws-cdk/aws-cloudfront";
 
 /**
  * This stack relies on getting the domain name from CDK context.
@@ -18,10 +21,31 @@ class TaylorsVersionSiteStack extends cdk.Stack {
     constructor(parent: cdk.App, name: string, props: cdk.StackProps) {
         super(parent, name, props);
 
-        new StaticSite(this, 'StaticSite', {
-            domainName: this.node.tryGetContext('domain'),
-            siteSubDomain: this.node.tryGetContext('subdomain'),
-        });
+        const domainName = this.node.tryGetContext('domain');
+        const siteSubDomain = this.node.tryGetContext('subdomain');
+
+        const siteDomain = siteSubDomain
+            ? siteSubDomain + '.' + domainName
+            : domainName;
+        new cdk.CfnOutput(this, 'Site', {value: 'https://' + siteDomain});
+
+        const {apiDomain} = new Notifications(this, "Backend")
+
+        new StaticDeployment(this, "Distribution", {
+            zoneName: domainName,
+            siteFqdn: siteDomain,
+            contentSource: [s3deploy.Source.asset('../dist')],
+            additionalBehaviors: [
+                {
+                    pathPattern: "/prod/notifications/*",
+                    origin: new origins.HttpOrigin(apiDomain),
+                    behaviorOptions: {
+                        viewerProtocolPolicy: cloudfront.ViewerProtocolPolicy.HTTPS_ONLY,
+                        allowedMethods: cloudfront.AllowedMethods.ALLOW_ALL
+                    }
+                }
+            ]
+        })
     }
 }
 
@@ -33,7 +57,7 @@ const prodStack = new TaylorsVersionSiteStack(app, 'TaylorsVersionProd', {
         account: process.env.CDK_DEPLOY_ACCOUNT
     }
 });
-Tags.of(prodStack).add('project', "taylors-version")
+cdk.Tags.of(prodStack).add('project', "taylors-version")
 
 const stagingStack = new TaylorsVersionSiteStack(app, 'TaylorsVersionStaging', {
     env: {
@@ -41,6 +65,14 @@ const stagingStack = new TaylorsVersionSiteStack(app, 'TaylorsVersionStaging', {
         account: process.env.CDK_DEPLOY_ACCOUNT
     }
 });
-Tags.of(stagingStack).add('project', "taylors-version")
+cdk.Tags.of(stagingStack).add('project', "taylors-version")
+
+const adhocStack = new TaylorsVersionSiteStack(app, 'TaylorsVersionAdhoc', {
+    env: {
+        region: process.env.CDK_DEPLOY_REGION,
+        account: process.env.CDK_DEPLOY_ACCOUNT
+    }
+});
+cdk.Tags.of(adhocStack).add('project', "taylors-version")
 
 app.synth();
