@@ -1,17 +1,10 @@
 import i18n from "i18next";
-import detector from "i18next-browser-languagedetector";
 import {initReactI18next} from "react-i18next";
 import en from "./locales/en.json";
-import fr from "./locales/fr.json";
 
-const resources = {
-    en,
-    fr, // french translations provided by https://www.reddit.com/user/perlimpinpute
-};
+const resources = {en};
 
 i18n
-    // pass in language detector
-    .use(detector)
     // pass the locales instance to react-i18next.
     .use(initReactI18next)
     // init i18next
