@@ -1,14 +1,21 @@
 import {CssBaseline, ThemeProvider} from "@material-ui/core";
 import React from "react";
-import LandingPage from "./components/LandingPage";
+import {Provider} from "react-redux";
+import {createStore} from "redux";
+import LandingLayout from "./components/LandingLayout";
+import rootReducer from "./reducers/rootReducer";
 import theme from "./themes/theme";
+
+const store = createStore(rootReducer);
 
 function Landing() {
     return (
-        <ThemeProvider theme={theme}>
-            <CssBaseline/>
-            <LandingPage/>
-        </ThemeProvider>
+        <Provider store={store}>
+            <ThemeProvider theme={theme}>
+                <CssBaseline/>
+                <LandingLayout/>
+            </ThemeProvider>
+        </Provider>
     );
 }
 
