@@ -1,3 +1,5 @@
+import fearless from "./fearless";
+
 const spotify = [
     // Single - Wildest Dreams
     {
@@ -17,409 +19,8 @@ const spotify = [
             "spotify:track:3ScVaMDQwaD8iumBYzcDvr",
         ],
     },
-
-    // Album - Fearless
-    // Fearless
-    {
-        "replacement": "spotify:track:77sMIMlNaSURUAXq5coCxE",
-        "originals": [
-            // original
-            "spotify:track:6Eu31gddWw0gOGO506pJYA",
-            // original extended
-            "spotify:track:61v7awvaSeBXJHBKMjw59I",
-            // original Canada
-            "spotify:track:3SBGfIPUVlqt1YGxgZXoJ9",
-            // BMRRS
-            "spotify:track:15PRxID4aGtOncbffzgZXI",
-            // (PE)
-            "spotify:track:2okho7vU7Nsq1UZD0kgIMi",
-            // (PE) unavailable
-            "spotify:track:4adub9Gizi8N2zCjq7XSpD",
-            // PE
-            "spotify:track:3F6czr26ZwGU5O5CHY04Ma",
-            // IV
-            "spotify:track:6zV8AVM9eVfdG7Qahwd64R",
-        ],
-    },
-
-    // Fifteen
-    {
-        "replacement": "spotify:track:2nqio0SfWg6gh2eCtfuMa5",
-        "originals": [
-            // original
-            "spotify:track:4t0OI7XrODjSkAu3bTPmWj",
-            // original extended
-            "spotify:track:07gxAwYjBbZdhmkxpMM0GP",
-            // original Canada
-            "spotify:track:3MXzl6hq3vSqB8ZlKVGkRi",
-            // BMRRS
-            "spotify:track:6aQrB8kaWyaCj9RYB7Rh5L",
-            // (PE)
-            "spotify:track:5Ora3Ish2A1JN5veGYTxhw",
-            // (PE) unavailable
-            "spotify:track:540uR4mP2DbmzOO4ydXsRX",
-            // PE
-            "spotify:track:3M0w3fhdXByHLCNAoi5c9G",
-            // IV
-            "spotify:track:4yf9OivCH2lmQWN8mzpCC6",
-            // rep STSSP
-            "spotify:track:0H2UD0LfSCiRr59YMGlCuS",
-        ],
-    },
-
-    // Love Story
-    {
-        "replacement": "spotify:track:3CeCwYWvdfXbZLXFhBrbnf",
-        "originals": [
-            // original
-            "spotify:track:1vrd6UOGamcKNGnSHJQlSt",
-            // original extended
-            "spotify:track:6K1sUckiJLXhsyte2yNFTQ",
-            // original Canada
-            "spotify:track:7bad461siaxECNlaYsmIo6",
-            // BMRRS
-            "spotify:track:1D4PL9B8gOg78jiHg3FvBb",
-            // (PE)
-            "spotify:track:5hnJCRSo4pxRsaEwT9YusG",
-            // (PE) unavailable
-            "spotify:track:4ikR3lDFcZ0hNKyVONfi4N",
-            // PE
-            "spotify:track:0kN3oXYWWAk1uC0y2WoyOE",
-            // IV
-            "spotify:track:5xPsybHWRHbTYO44xnREY5",
-        ],
-    },
-
-    // Hey Stephen
-    {
-        "replacement": "spotify:track:550erGcdD9n6PnwxrvYqZT",
-        "originals": [
-            // original
-            "spotify:track:4WXzzCof26KJLTK5kK53dS",
-            // original extended
-            "spotify:track:5SyE6zNX1wSRobzxNkeX43",
-            // original Canada
-            "spotify:track:4T6pHxntdd0SZrjKtt8nfM",
-            // BMRRS
-            "spotify:track:3DTqKBlJPlRSclFfBI0ZoE",
-            // (PE)
-            "spotify:track:3CI1f4yLwdHzcUnH7o0Xyh",
-            // (PE) unavailable
-            "spotify:track:32pM2R8GQLtbBKdSwdTAM1",
-            // PE
-            "spotify:track:6wcFdMmiDscAhZJrcVFumx",
-            // IV
-            "spotify:track:4fWDgRZa0GKE5KZNhE9MQa",
-            // rep STSSP
-            "spotify:track:0TYJbIHBtmWy9ugtU9BcbE",
-        ],
-    },
-
-    // White Horse
-    {
-        "replacement": "spotify:track:5YL553x8sHderRBDlm3NM3",
-        "originals": [
-            // original
-            "spotify:track:6wn61Fzx9XMxQmieLpoIhW",
-            // original extended
-            "spotify:track:2OGA4hbUqiAhCQj1pxeNld",
-            // original Canada
-            "spotify:track:5GEPI2EfHqf6G1Ebibo5C5",
-            // BMRRS
-            "spotify:track:4QhjEJ2o6QNwpZGs51AmiY",
-            // (PE)
-            "spotify:track:5yKB8qcTSflDercM2PXwot",
-            // (PE) unavailable
-            "spotify:track:1vicNVzWoAnteQMlTzI2VF",
-            // PE
-            "spotify:track:4jPtq9shRdUv8Zjr0a5xkD",
-            // IV
-            "spotify:track:1yFjBRQVhRdcdnrTE5LuyD",
-            // rep STSSP
-            "spotify:track:4RMfSYDFkcz0bKoK8fHDgR",
-        ],
-    },
-
-    // You Belong With Me
-    {
-        "replacement": "spotify:track:1qrpoAMXodY6895hGKoUpA",
-        "originals": [
-            // original
-            "spotify:track:3GCL1PydwsLodcpv0Ll1ch",
-            // original extended
-            "spotify:track:6mqdunuFFSODHKcpDTFvAj",
-            // original Canada
-            "spotify:track:3OaBCFROgrLRbZGCFP6E3M",
-            // BMRRS
-            "spotify:track:1GEBsLDvJGw7kviySRI6GX",
-            // (PE)
-            "spotify:track:4kgTdThcDHfuDS2kKxB7Lc",
-            // (PE) unavailable
-            "spotify:track:26eccs3bbw6DMekFwZbdL2",
-            // PE
-            "spotify:track:3as6Wz3mwbTnusWhZjTUNr",
-            // IV
-            "spotify:track:7G6TDmGJXDoBaV6NGj3Yic",
-        ],
-    },
-
-    // Breathe
-    {
-        "replacement": "spotify:track:7HC7R2D8WjXVcUHJyEGjRs",
-        "originals": [
-            // original
-            "spotify:track:49mWEy5MgtNujgT7xU3emT",
-            // original extended
-            "spotify:track:3uvCE2wdHDyP0AojlmPlUa",
-            // original Canada
-            "spotify:track:0ykAwrzigpz5BACbaUyDm3",
-            // BMRRS
-            "spotify:track:2dzk2z7MXl0peRhBUsDq4d",
-            // (PE)
-            "spotify:track:23qIWGhpwvDgC23TyZOUck",
-            // (PE) unavailable
-            "spotify:track:20h5sJ9pElnXT4cbVUTkHs",
-            // PE
-            "spotify:track:6S9G7J3EB1ZY0rrZPwsBg5",
-            // IV
-            "spotify:track:0qsSmObOgBZfuMIwfAQDBP",
-            // rep STSSP
-            "spotify:track:3oxgFdIlxFEDEJqn7k5TVN",
-        ],
-    },
-
-    // Tell Me Why
-    {
-        "replacement": "spotify:track:0k0vFacOHNuArLWMiH60p7",
-        "originals": [
-            // original
-            "spotify:track:3rnI1UCyGJvUTVvT97VQr5",
-            // original extended
-            "spotify:track:5du2fre21jWENkp1IDRaXf",
-            // original Canada
-            "spotify:track:4xUskJSjCmC0kSi3JmjiIz",
-            // BMRRS
-            "spotify:track:7bUvIh9rQsboJBXaSIQZAF",
-            // (PE)
-            "spotify:track:4x7Gxk2GfwZJcqpoVT1PbR",
-            // (PE) unavailable
-            "spotify:track:6DfeQ1pT4ClDJKLeQOTnKW",
-            // PE
-            "spotify:track:4XrP7vKBjKLEYlRZdb0Ygr",
-            // IV
-            "spotify:track:6vLpy7UPYuOa75Bkqocrcw",
-        ],
-    },
-
-    // You're Not Sorry
-    {
-        "replacement": "spotify:track:6iiAfo4wTA2CVC3Uwx9uh8",
-        "originals": [
-            // original
-            "spotify:track:0HmCuN0Z3OX1Qrz43FLOPL",
-            // original extended
-            "spotify:track:5SSYD3BoD18792DOlW0h5V",
-            // original Canada
-            "spotify:track:5kCqMXvv33WEcpb3Xvx1FZ",
-            // BMRRS
-            "spotify:track:0ReNbPVMU7jUWqn1yHo9cE",
-            // (PE)
-            "spotify:track:3JhvYnOfKLJuWC7jEU0GmJ",
-            // (PE) unavailable
-            "spotify:track:7zen3j1316Tsk5iEZuDOft",
-            // PE
-            "spotify:track:6HWYtS215rxaaMjvpyG18W",
-            // IV
-            "spotify:track:4N419mmxYybTfAECy5j4dw",
-        ],
-    },
-
-    // The Way I Loved You
-    {
-        "replacement": "spotify:track:22bPsP2jCgbLUvh82U0Z3M",
-        "originals": [
-            // original
-            "spotify:track:5P4wWhUYWM0IaVYLuZxdar",
-            // original extended
-            "spotify:track:6MNPOhGO5dv15nqei7GN0D",
-            // original Canada
-            "spotify:track:4ONlmL2EYEryaJzmh9fG1o",
-            // BMRRS
-            "spotify:track:54ziHHVIBJg3YJ9sn17mEy",
-            // (PE)
-            "spotify:track:4pEZUytwBzha5Un6m3SCFf",
-            // (PE) unavailable
-            "spotify:track:7h8gWtGOVLLpgOvVYNFqC7",
-            // PE
-            "spotify:track:77VotUUL79U10TAnYdqwvr",
-            // IV
-            "spotify:track:04LDcz3RX1P5sIlzQGqgUN",
-        ],
-    },
-
-    // Forever & Always
-    {
-        "replacement": "spotify:track:1msEuwSBneBKpVCZQcFTsU",
-        "originals": [
-            // original
-            "spotify:track:47HtKpfzpAt8rQjjXWotFj",
-            // original extended
-            "spotify:track:3wxkYiniEnp0ZafDE5vD07",
-            // original Canada
-            "spotify:track:30oulfBS2vZcq1wQZPXlQo",
-            // BMRRS
-            "spotify:track:5VHdgl9hFUmabiCO5jeISv",
-            // (PE)
-            "spotify:track:3NKCWBT2uKp8rxSkcvIrdX",
-            // (PE) unavailable
-            "spotify:track:4oYMFlaSqdoID82aQtXPYg",
-            // PE
-            "spotify:track:1zxrcAk6eiytfavqriMcKT",
-            // IV
-            "spotify:track:3yuOL0jFnVR8tizZzNNo0F",
-            // rep STSSP
-            "spotify:track:0QCS4NpURHvWdF1icdkfBw",
-        ],
-    },
-
-    // The Best Day
-    {
-        "replacement": "spotify:track:6ON9UuIq49xXY9GPmHIYRp",
-        "originals": [
-            // original
-            "spotify:track:3esA216TyLHEkNiBCeCmcg",
-            // original extended
-            "spotify:track:4G5YAPiKudJCDmICzj0Zwk",
-            // original Canada
-            "spotify:track:6LAIapZk9Ozzhau5gBDNmi",
-            // BMRRS
-            "spotify:track:0ck1qErO76fM9xKxuSeqDk",
-            // (PE)
-            "spotify:track:7CjuAoOiSrwWv7wJonAUmw",
-            // (PE) unavailable
-            "spotify:track:0U04szAxhhjbqHgwy9ndoh",
-            // PE
-            "spotify:track:5YbZZ2gYfvW1UvHHF4pVaD",
-            // IV
-            "spotify:track:7G9D32Kx7hdQ9Uga1V4VXm",
-        ],
-    },
-
-    // Change
-    {
-        "replacement": "spotify:track:3ExweHKZF9B752DPQByRVT",
-        "originals": [
-            // original
-            "spotify:track:1yACRKAwlXWhXXFUSkvzhD",
-            // original extended
-            "spotify:track:6rEdUI8O9sPKbN1HQKnbrL",
-            // original Canada
-            "spotify:track:4BuxFaXa4RblrdN57lWN6d",
-            // BMRRS
-            "spotify:track:6nIUWFgMly8cdh32SbnhiR",
-            // (PE)
-            "spotify:track:3gFaNwyiafgUoQhR0pHYA9",
-            // (PE) unavailable
-            "spotify:track:4EnxFZptr8J9Ytzz7uyElf",
-            // PE
-            "spotify:track:0RdcBnAXyH2YZZvcBVu2k6",
-            // IV
-            "spotify:track:493jXNrYmXPCMbSJvVRYLk",
-            // rep STSSP
-            "spotify:track:5PSOWHll3h3Vy226xSJk0M",
-        ],
-    },
-
-    // Jump Then Fall
-    {
-        "replacement": "spotify:track:2m3ObD945KvpE5y9A1eUWm",
-        "originals": [
-            // (PE)
-            "spotify:track:6WWvcA62QlnRguluWAk6bv",
-            // (PE) unavailable
-            "spotify:track:4mvQXr3EeckZIBmlcxLqXe",
-            // PE
-            "spotify:track:08gavXombT6KR0af88i9tA",
-            // rep STSSP
-            "spotify:track:5zytSTR2g0I9psX2Z12ex6",
-        ],
-    },
-
-    // Untouchable
-    {
-        "replacement": "spotify:track:0tQ9vBYpldCuikPsbgOVKA",
-        "originals": [
-            // (PE)
-            "spotify:track:6ovFJkdau79DuXEvQjPaNH",
-            // (PE) unavailable
-            "spotify:track:5IvkGi34lJuiZp9gdlafJ1",
-            // PE
-            "spotify:track:2IZ00ed83ygPIiacYScWUE",
-        ],
-    },
-
-    // Forever & Always (Piano Version)
-    {
-        "replacement": "spotify:track:01QdEx6kFr78ZejhQtWR5m",
-        "originals": [
-            // (PE)
-            "spotify:track:13MAWVUSkbGMUi6I7RowRK",
-            // (PE) unavailable
-            "spotify:track:1QK0cJlXnfq3q7Vo8mVL4w",
-            // PE
-            "spotify:track:46HGgtwmmuEB8mvDCyjyAc",
-        ],
-    },
-
-    // Come In With The Rain
-    {
-        "replacement": "spotify:track:1n2wszmJyVkw6FHqyLnQsY",
-        "originals": [
-            // (PE)
-            "spotify:track:0cEG04R5cldkTLzK6phlcn",
-            // (PE) unavailable
-            "spotify:track:4xLiykjcfrGMkzsHHK4Lhh",
-            // PE
-            "spotify:track:4pl5zcqCv4vc3cB7M4MZ6f",
-        ],
-    },
-
-    // SuperStar
-    {
-        "replacement": "spotify:track:51A8eKvvZz9uydvIZ7xRSV",
-        "originals": [
-            // (PE)
-            "spotify:track:0XfTh16YcKPzdxMPTeKTAq",
-            // (PE) unavailable
-            "spotify:track:313c46YL18vX1rKsyF5oc6",
-            // PE
-            "spotify:track:14Bljc3pOOG0xQX3wqhLN9",
-        ],
-    },
-
-    // The Other Side Of The Door
-    {
-        "replacement": "spotify:track:1cSFlSBdpT4F5vb1frQ231",
-        "originals": [
-            // (PE)
-            "spotify:track:1yDJcXc549N6AabNBzFAku",
-            // (PE) unavailable
-            "spotify:track:7JbxkpFLB6pvuNu3buoxAj",
-            // PE
-            "spotify:track:0xvsgzM8AtBtRHZm5rav8A",
-        ],
-    },
-
-    // Today Was A Fairytale
-    {
-        "replacement": "spotify:track:2JoJrsEV15OzbijS47lids",
-        "originals": [
-            // Single
-            "spotify:track:4pFvEWbjBpPUdYRQly0THs",
-            // US Version (Valentine's Day Cover Art)
-            "spotify:track:1P5QpwOgXJuoVTm3CWbnLY",
-        ],
-    },
+    ...fearless,
+    //...red
 ];
 
 export default spotify;
